refactor logic folder
