+++ packages/ui/src/components/index.ts
@@ -17,7 +17,6 @@
  */
 
 export * from './atoms';
-export * from './dashboard';
 export * from './molecules';
 export * from './organisms';
 export * from './templates';
