+++ packages/ui/src/components/templates/Dashboard/chart.ts
