+++ packages/ui/src/components/templates/index.js
@@ -16,5 +16,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+export {default as Dashboard} from './Dashboard/Dashboard';
+export {default as BarChart} from './Dashboard/Chart/BarChart';
+export {default as LineChart} from './Dashboard/Chart/LineChart';
+export {default as PieChart} from './Dashboard/Chart/PieChart';
 export {default as ObjectCard} from './ObjectCard/ObjectCard';
 export {default as TabsScreen} from './TabsScreen/TabsScreen';
